Add news Features and enhanced desing

+++ SoreBooksBlazorWASM.Service/VentasManagerV1.cs
@@ -15,26 +15,18 @@ namespace SoreBooksBlazorWASM.Service
     public class VentasManagerV1
     {
         private readonly ApplicationDbContext _context;
-        private readonly AuthenticationStateProvider _authProvider;
-        private readonly UserManager<IdentityUser> _userManager;
 
-        public VentasManagerV1(ApplicationDbContext context, AuthenticationStateProvider authProvider, UserManager<IdentityUser> userManager)
+        public VentasManagerV1( ApplicationDbContext context)
         {
             _context = context;
-            _authProvider = authProvider;
-            _userManager = userManager;
         }
 
-        public async Task CompletarVenta(VentaViewModel model)
+        public async Task CompletarVenta(VentaViewModel model, string id)
         {
             try
             {
-                var authState = await _authProvider.GetAuthenticationStateAsync();
-                var user = authState.User;
-                var usuarioActual = await _userManager.GetUserAsync(user);
-
                 var buscarVentaActiva = await _context.Ventas
-                    .Where(u => u.IdUsuario == usuarioActual.Id && u.Estado == "En Proceso")
+                    .Where(u => u.IdUsuario == id && u.Estado == "En Proceso")
                     .FirstOrDefaultAsync();
 
 
@@ -43,128 +35,182 @@ namespace SoreBooksBlazorWASM.Service
                     throw new Exception("Carrito de Compras Vacio");
                 }
 
-                var TotalVenta = await _context.DetalleVentas
-                    .Where(dv => dv.IdVenta == buscarVentaActiva.IdVenta)
-                    .SumAsync(d => d.Cantidad);
 
                 buscarVentaActiva.Estado = "Finalizada";
                 buscarVentaActiva.MetodoPago = model.MetodoPago;
                 buscarVentaActiva.DireccionEnvio = model.DireccionEnvio;
-                buscarVentaActiva.Total = TotalVenta;
 
                 await _context.SaveChangesAsync();
             }
             catch(Exception ex) 
             {
-                throw;
+                throw new Exception(ex.Message);
             }
         }
 
-        public async Task AgregarCarrito(LibroViewModel model)
-        {
-            var authState = await _authProvider.GetAuthenticationStateAsync();
-            var user = authState.User;
-            var usuarioActual = await _userManager.GetUserAsync(user);
 
-            var buscarVentaActiva = await _context.Ventas
-                .Where(u => u.IdUsuario == usuarioActual.Id && u.Estado == "En Proceso")
-                .FirstOrDefaultAsync();
-
-            if (buscarVentaActiva == null)
-            {
-                var nuevaVenta = new Ventas
-                {
-                    IdUsuario = usuarioActual.Id,
-                    FechaVenta = DateTime.Now,
-                    Estado = "En Proceso"
-                };
-
-                _context.Ventas.Add(nuevaVenta);
-                await _context.SaveChangesAsync(); 
-
-                buscarVentaActiva = nuevaVenta;
-            }
-
-            var detalleExistente = await _context.DetalleVentas
-                .Where(dv => dv.IdVenta == buscarVentaActiva.IdVenta && dv.IdLibro == model.IdLibro)
-                .FirstOrDefaultAsync();
-
-            if (detalleExistente != null)
-            {
-                detalleExistente.Cantidad += 1;
-                detalleExistente.TotalVenta = detalleExistente.Cantidad * detalleExistente.PrecioUnitario;
-
-            }
-            else
-            {
-                var nuevoDetalle = new DetalleVentas
-                {
-                    IdVenta = buscarVentaActiva.IdVenta,
-                    IdLibro = model.IdLibro,
-                    Cantidad = 1,
-                    PrecioUnitario = model.Precio,
-                    TotalVenta = model.Precio 
-                };
-
-                detalleExistente = nuevoDetalle;
-
-                _context.DetalleVentas.Add(nuevoDetalle);
-            }
-
-            buscarVentaActiva.Total += detalleExistente.TotalVenta; 
-
-            await _context.SaveChangesAsync();
-        }
-
-        public async Task<List<VentaGeneral>> DetalleGeneralVenta()
+        public async Task<VentaViewModel> ObtenerVenta(string userid)
         {
             try
             {
-                var authState = await _authProvider.GetAuthenticationStateAsync();
-                var user = authState.User;
-                var usuarioActual = await _userManager.GetUserAsync(user);
+                var buscarVenta = await _context.Ventas
+                    .Where(v => v.IdUsuario.Equals(userid) && v.Estado.Equals("En Proceso"))
+                    .Select(x => new VentaViewModel
+                    {
+                        IdVenta = x.IdVenta,
+                        IdUsuario = x.IdUsuario,
+                        Estado = x.Estado,
+                        FechaVenta = x.FechaVenta,
+                        MetodoPago = x.MetodoPago,
+                        DireccionEnvio=x.DireccionEnvio,
+                        Total = x.Total
+                    }).FirstOrDefaultAsync();
 
-                List<VentaGeneral> listaVentaGeneral = new List<VentaGeneral>();
+                return buscarVenta;
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
+        }
 
+        public async Task AgregarCarrito(LibroViewModel model, string id)
+        {
+            try
+            {
                 var buscarVentaActiva = await _context.Ventas
-                    .Where(u => u.IdUsuario == usuarioActual.Id && u.Estado == "En Proceso")
+                    .Where(u => u.IdUsuario == id && u.Estado == "En Proceso")
                     .FirstOrDefaultAsync();
-
 
                 if (buscarVentaActiva == null)
                 {
-                    throw new Exception("No tiene articulos añadidos");
+                    var nuevaVenta = new Ventas
+                    {
+                        IdUsuario = id,
+                        FechaVenta = DateTime.Now,
+                        Estado = "En Proceso"
+                    };
+
+                    _context.Ventas.Add(nuevaVenta);
+                    await _context.SaveChangesAsync();
+
+                    buscarVentaActiva = nuevaVenta;
                 }
 
+                var detalleExistente = await _context.DetalleVentas
+                    .Where(dv => dv.IdVenta == buscarVentaActiva.IdVenta && dv.IdLibro == model.IdLibro)
+                    .FirstOrDefaultAsync();
+
+                if (detalleExistente != null)
+                {
+                    detalleExistente.Cantidad += 1;
+                    detalleExistente.TotalVenta = detalleExistente.Cantidad * detalleExistente.PrecioUnitario;
+
+                }
+                else
+                {
+                    var nuevoDetalle = new DetalleVentas
+                    {
+                        IdVenta = buscarVentaActiva.IdVenta,
+                        IdLibro = model.IdLibro,
+                        Cantidad = 1,
+                        PrecioUnitario = model.Precio,
+                        TotalVenta = model.Precio
+                    };
+
+                    detalleExistente = nuevoDetalle;
+
+                    _context.DetalleVentas.Add(nuevoDetalle);
+                }
+
+                var sumaTotalDetalles = await _context.DetalleVentas
+                    .Where(dv => dv.IdVenta == buscarVentaActiva.IdVenta && dv.IdLibro == model.IdLibro)
+                    .SumAsync(dv => dv.TotalVenta);
+
+                buscarVentaActiva.Total += sumaTotalDetalles;
+
+                await _context.SaveChangesAsync();
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
+        }
+
+        public async Task<List<VentaGeneral>> DetalleGeneralVenta(string id)
+        {
+            try
+            {
+                List<VentaGeneral> listaVentaGeneral = new List<VentaGeneral>();
+
+                // Buscar la venta activa
+                var buscarVentaActiva = await _context.Ventas
+                    .Where(u => u.IdUsuario == id && u.Estado == "En Proceso")
+                    .FirstOrDefaultAsync();
+
+                // Si no se encuentra la venta activa, lanzar una excepción
+                if (buscarVentaActiva == null)
+                {
+                    throw new Exception("No tiene artículos añadidos o la venta no está en proceso.");
+                }
+
+                // Obtener detalles de venta
                 var buscarDetalleVenta = await _context.DetalleVentas
                     .Where(dv => dv.IdVenta == buscarVentaActiva.IdVenta)
                     .ToListAsync();
 
-                foreach (var item in buscarDetalleVenta)
+                if (buscarDetalleVenta == null)
                 {
-                    var nuevoDetalle = new VentaGeneral
-                    {
-                        IdDetalleVentas = item.IdDetalleVentas,
-                        IdLibro = item.IdLibro,
-                        IdVenta = item.IdVenta,
-                        ImgLibro = item.Libro.Imagen,
-                        Titulo = item.Libro.Titulo,
-                        Cantidad = item.Cantidad,
-                        PrecioUnitario = item.PrecioUnitario,
-                        TotalVenta = item.Ventas.Total,
-                    };
-
-                    listaVentaGeneral.Add(nuevoDetalle);
+                    throw new Exception("No tiene artículos añadidos");
                 }
 
-                return listaVentaGeneral;
+                // Obtener los libros en una sola consulta para evitar múltiples consultas
+                var idsLibros = buscarDetalleVenta.Select(dv => dv.IdLibro).Distinct().ToList();
 
+                var libros = await _context.Libros
+                    .Where(l => idsLibros.Contains(l.IdLibro))
+                    .ToDictionaryAsync(l => l.IdLibro);
+
+                foreach (var item in buscarDetalleVenta)
+                {
+                    // Buscar el libro correspondiente usando el diccionario
+                    if (libros.TryGetValue(item.IdLibro, out var buscarLibro))
+                    {
+                        var nuevoDetalle = new VentaGeneral
+                        {
+                            IdDetalleVentas = item.IdDetalleVentas,
+                            IdLibro = item.IdLibro,
+                            IdVenta = item.IdVenta,
+                            ImgLibro = buscarLibro.Imagen,
+                            Titulo = buscarLibro.Titulo,
+                            Cantidad = item.Cantidad,
+                            PrecioUnitario = item.PrecioUnitario,
+                            TotalVenta = item.TotalVenta
+                        };
+
+                        listaVentaGeneral.Add(nuevoDetalle);
+                    }
+                    else
+                    {
+                        // Manejar el caso en que no se encuentra el libro
+                        // Por ejemplo, puedes registrar un error o agregar un mensaje de advertencia
+                    }
+                }
+
+                var sumaTotalVenta = buscarDetalleVenta.Sum(dv => dv.TotalVenta);
+                buscarVentaActiva.Total = sumaTotalVenta;
+
+                await _context.SaveChangesAsync();
+
+                return listaVentaGeneral;
             }
             catch (Exception ex)
             {
-                throw;
+                // Manejar la excepción adecuadamente, por ejemplo, registrando el error
+                throw new HttpRequestException("Error al obtener los detalles de la venta, " + ex.Message);
             }
         }
+
 
 
         public async Task EliminarProductoDetalle(int idDetalleVenta)
@@ -175,6 +221,10 @@ namespace SoreBooksBlazorWASM.Service
                     .Where(dv => dv.IdDetalleVentas == idDetalleVenta)
                     .FirstOrDefaultAsync();
 
+                var contarDetalles = await _context.DetalleVentas
+                    .Select(x => x.IdDetalleVentas)
+                    .CountAsync();
+
                 var buscarVenta = await _context.Ventas
                     .Where(v => v.IdVenta == buscarDetalle.IdVenta)
                     .FirstOrDefaultAsync();
@@ -182,7 +232,12 @@ namespace SoreBooksBlazorWASM.Service
                 if (buscarDetalle.Cantidad == 1)
                 {
                     _context.DetalleVentas.Remove(buscarDetalle);
-                    _context.Ventas.Remove(buscarVenta);
+
+                    if (contarDetalles == 1)
+                    {
+                        _context.Ventas.Remove(buscarVenta);
+                    }
+
                 }
                 else if(buscarDetalle.Cantidad >1)
                 {
@@ -195,13 +250,15 @@ namespace SoreBooksBlazorWASM.Service
 
                 }
 
+                
+
                 await _context.SaveChangesAsync();                
 
 
             }
             catch (Exception ex)
             {
-                throw;
+                throw new Exception(ex.Message);
             }
         }
